KafkaToInfluxApp/requirements.txt
kafka-python
influxdb